fix(render): allow the frame rates editing timelines actually use
The overlay started almost in sync in Final Cut and fell further behind as the
song went on. The rendered file was not at fault: 4495 frames, a constant 25/1,
179.800s, starting at zero. What was at fault is that 25 was as close as it
could get.

fps was an integer, so 23.976 and 29.97 were refused outright — and those are
the rates most timelines run at. A clip at a rate the project does not use gets
conformed, and a conform is exactly this symptom: 25 stretched to 24 is seven
seconds late by three minutes in.

Rates are fractions now, and the NTSC ones are kept as fractions rather than the
decimals people write them as. 23.976 encodes as 24000/1001, and frame times are
computed from that fraction too — the decimal would put the same drift back,
about two frames an hour, which is the kind of thing that gets found much later.

The transparent overlay preset also carries the song now. A clip holding its own
audio lines itself up in an editor and cannot drift away from it, which is worth
more than not having a track to mute. Turned off with --no-audio or the checkbox
in the export dialog, which also gained the NTSC rates.

+++ src/videokar/web/app.py
@@ -58,7 +58,7 @@ class RenderRequest(BaseModel):
     format: str | None = None
     width: int | None = None
     height: int | None = None
-    fps: int | None = None
+    fps: float | None = None
     overrides: dict[str, Any] | None = None
     """Anything else from the style schema. Validated by resolve_style, so the
     page can offer a control for a setting without the server learning its name."""
